Centralized the Principal Types in a higher-up directory.
Some types are so common throughout the app, they deserve to be central.
