+++ src/mfa/totp/types.ts
@@ -1,4 +1,4 @@
-import { User } from '../../principal/types';
+import { User } from '../../types';
 
 export type TotpDevice = {
     user: User;
