+++ src/oauth2-client/formats/hal.ts
@@ -1,4 +1,4 @@
-import { App } from '../../principal/types';
+import { App } from '../../types';
 import { OAuth2Client } from '../types';
 import { HalResource } from 'hal-types';
