+++ src/reset-password/controller/reset-password.ts
@@ -4,7 +4,7 @@ import { Forbidden } from '@curveball/http-errors';
 import log from '../../log/service';
 import { EventType } from '../../log/types';
 import * as UserService from '../../user/service';
-import { User } from '../../principal/types';
+import { User } from '../../types';
 import { resetPasswordForm } from '../formats/redirect';
 
 class ResetPasswordController extends Controller {
