+++ src/types.ts
@@ -18,3 +18,44 @@ export type ServerStats = {
   tokensIssued: number;
 }
 
+/**
+ * List of possible principal types.
+ */
+export type PrincipalType = 'user' | 'app' | 'group';
+
+export const PrincipalTypeList: PrincipalType[] = ['user', 'app', 'group'];
+
+/** Base principal that everything inherits from
+ */
+export type BasePrincipal<TType extends PrincipalType> = {
+  id: number;
+  href: string;
+  externalId: string;
+  type: TType;
+  identity: string;
+  nickname: string;
+  createdAt: Date;
+  modifiedAt: Date;
+  active: boolean;
+}
+
+/**
+ * User. This represents a person
+ */
+export type User = BasePrincipal<'user'>;
+
+/**
+ * A group can contain multiple users or other groups and lets you assign
+ * privileges to entire groups of people.
+ */
+export type Group = BasePrincipal<'group'>;
+
+/**
+ * Apps represent systems / oauth2 clients.
+ */
+export type App = BasePrincipal<'app'>;
+
+/**
+ * Any principal
+ */
+export type Principal = User | Group | App;
